models/A2C/model_requirements.txt
torch>=1.9.0
numpy>=1.21.0
requests>=2.25.0
scikit-learn>=1.0.0
nltk>=3.6.0
vaderSentiment>=3.3.2
transformers>=4.41.0
sentence-transformers>=2.2.0
matplotlib>=3.5.0
seaborn>=0.11.0
tqdm>=4.62.0
pandas>=1.3.0
python-dotenv>=0.19.0
PyYAML>=5.4.0
wikipedia>=1.4.0
flask>=2.0.0 